+++ src/main/java/org/j2c/llvm/LLVM.java
@@ -1,6 +1,6 @@
 package org.j2c.llvm;
 
-import org.j2c.assembly.NClass;
+import org.j2c.ast.NClass;
 
 public class LLVM {
     static {
