+++ src/main/kotlin/org/j2c/ast/rules/arrays.kt
@@ -1,7 +1,9 @@
-package org.j2c.assembly.rules
+package org.j2c.ast.rules
 
 import javassist.bytecode.Opcode
-import org.j2c.assembly.NArrayLength
+import org.j2c.ast.NArrayLength
+import org.j2c.ast.rules.api.Rule
+import org.j2c.ast.rules.api.RuleContainer
 
 @RuleContainer
 object ARRAYLENGTH {
