+++ src/main/kotlin/org/j2c/ast/rules/iload_istore_iconst.kt
@@ -1,9 +1,11 @@
-package org.j2c.assembly.rules
+package org.j2c.ast.rules
 
 import javassist.bytecode.Opcode
-import org.j2c.assembly.NAssignment
-import org.j2c.assembly.NInt
-import org.j2c.assembly.NReference
+import org.j2c.ast.NAssignment
+import org.j2c.ast.NInt
+import org.j2c.ast.NReference
+import org.j2c.ast.rules.api.Rule
+import org.j2c.ast.rules.api.RuleContainer
 
 @RuleContainer
 object ILOAD {
