+++ src/main/kotlin/org/j2c/ast/rules/types.kt
@@ -1,6 +1,8 @@
-package org.j2c.assembly.rules
+package org.j2c.ast.rules
 
 import javassist.bytecode.Opcode
+import org.j2c.ast.rules.api.Rule
+import org.j2c.ast.rules.api.RuleContainer
 
 @RuleContainer
 object CHECKCAST {
