+++ src/test/kotlin/org/j2c/lang_test.kt
@@ -1,7 +1,7 @@
 package org.j2c
 
-import org.j2c.assembly.rules.NoRule
-import org.j2c.assembly.rules.RuleContainer
+import org.j2c.ast.rules.api.NoRule
+import org.j2c.ast.rules.api.RuleContainer
 import org.junit.jupiter.api.Test
 import org.reflections.Reflections
 import org.reflections.scanners.SubTypesScanner
@@ -15,8 +15,8 @@ class LanguageTests {
     fun getRules() {
         val reflections = Reflections(
             ConfigurationBuilder()
-                .forPackages("org.j2c.assembly.rules")
-                .filterInputsBy(FilterBuilder().includePackage("org.j2c.assembly.rules"))
+                .forPackages("org.j2c.ast.rules")
+                .filterInputsBy(FilterBuilder().includePackage("org.j2c.ast.rules"))
                 .setScanners(SubTypesScanner(false))
         )
         val classes = reflections.getSubTypesOf(Any::class.java)
